% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sheets.R
\name{list_sheets}
\alias{list_sheets}
\title{List the sheets in an '.xlsb' workbook}
\usage{
list_sheets(path)
}
\arguments{
\item{path}{Path to the \code{.xlsb} file.}
}
\value{
A character vector of sheet names.
}
\description{
Returns the names of the worksheets in an Excel binary (\code{.xlsb}) workbook,
in workbook order. This is the \code{.xlsb} counterpart to
\code{\link[readxl:excel_sheets]{readxl::excel_sheets()}}.
}
\examples{
\dontrun{
list_sheets("workbook.xlsb")
}

}
\seealso{
\code{\link[=rxlsb]{rxlsb()}} to read a sheet.
}
